feat: move ant update logic into useAntKiller hook
